timers: Fix timer integer logic
fire_in and period is tightly coupled in timer logic
to fix observed integer overflow make types identical

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/55783
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>

+++ lib/timers/timers.c
@@ -36,7 +36,8 @@ spinlock_t tm_lock;
 
 static int num_started_timers = 0;
 
-static void hw_timer_set_period(uint32_t ticks) {
+static void hw_timer_set_period(uint64_t ticks)
+{
 	hw_timer_period = ticks;
 	plat_timer_set_period(ticks);
 }
